requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
line-bot-sdk==3.7.0
python-dotenv==1.0.0
psycopg2==2.9.9
alembic==1.13.1
geopy==2.4.1
